refactor: 전체 테스트에 fixture 적용
