+++ back/src/test/java/com/baba/back/baby/acceptance/BabyAcceptanceTest.java
@@ -1,5 +1,7 @@
 package com.baba.back.baby.acceptance;
 
+import static com.baba.back.fixture.DomainFixture.멤버1;
+import static com.baba.back.fixture.DomainFixture.아기1;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
@@ -25,8 +27,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 
 public class BabyAcceptanceTest extends AcceptanceTest {
-    public static final String MEMBER_ID = "1234";
-    public static final String BABY_ID = "1234";
+
     public static final String BABY_BASE_PATH = "/baby";
 
     @Autowired
@@ -51,7 +52,7 @@ public class BabyAcceptanceTest extends AcceptanceTest {
     @Test
     void 기본_설정된_아기가_없으면_404를_던진다() {
         // given
-        final String token = tokenProvider.createToken(MEMBER_ID);
+        final String token = tokenProvider.createToken(멤버1.getId());
 
         // when
         ExtractableResponse<Response> response = RestAssured.given()
@@ -69,27 +70,10 @@ public class BabyAcceptanceTest extends AcceptanceTest {
     @Test
     void 디폴트_아기를_조회한다() {
         // given
-        final String token = tokenProvider.createToken(MEMBER_ID);
+        Member member = memberRepository.save(멤버1);
+        final String token = tokenProvider.createToken(멤버1.getId());
 
-        LocalDate birthday = LocalDate.of(2024, 1, 25);
-        LocalDate now = LocalDate.of(2023, 1, 25);
-        final String color = "FFAEBA";
-        ColorPicker<String> colorPicker = (List<String> colors) -> color;
-
-        Member member = memberRepository.save(Member.builder()
-                .id(MEMBER_ID)
-                .name("박재희")
-                .introduction("")
-                .colorPicker(colorPicker)
-                .iconName("icon1")
-                .build());
-
-        Baby baby = babyRepository.save(Baby.builder()
-                .id(BABY_ID)
-                .name("앙쥬")
-                .birthday(birthday)
-                .now(now)
-                .build());
+        Baby baby = babyRepository.save(아기1);
 
         relationRepository.save(Relation.builder()
                 .member(member)
@@ -111,7 +95,7 @@ public class BabyAcceptanceTest extends AcceptanceTest {
         // then
         assertAll(
                 () -> assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value()),
-                () -> assertThat(response.response().jsonPath().getString("babyId")).isEqualTo(BABY_ID)
+                () -> assertThat(response.response().jsonPath().getString("babyId")).isEqualTo(baby.getId())
         );
 
     }
