+++ back/src/test/java/com/baba/back/oauth/service/MemberServiceTest.java
@@ -1,7 +1,7 @@
 package com.baba.back.oauth.service;
 
-import static com.baba.back.fixture.Fixture.이미_회원가입한_유저1;
-import static com.baba.back.fixture.Fixture.회원가입_안한_유저1;
+import static com.baba.back.fixture.DomainFixture.이미_회원가입한_유저1;
+import static com.baba.back.fixture.DomainFixture.회원가입_안한_유저1;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.mockito.ArgumentMatchers.any;
